Corrigiendo PR
