+++ src/function/pass.ts
@@ -1,0 +1,12 @@
+import LeerArchivo from './readfile';
+
+const path = "password.txt"
+
+const getPassword = () => {
+    return LeerArchivo(path);
+}
+
+export default getPassword;
+
+
+
