+++ src/index.ts
@@ -4,8 +4,6 @@ import express from "express";
 import fs from "fs";
 import schema from "./schema.json";
 import server from "./server";
-import {listProfile ,createProfile} from "./client";
- 
 
 export const router = express.Router({
   strict: true,
@@ -92,9 +90,3 @@ server.post("/", async (req: Request, res: Response) => {
 server.listen("80", () => {
   console.log("listening");
 });
-
-const ejecutar_cliente = () => {
-console.log(createProfile());
-};
-
-ejecutar_cliente();
